Handle the case of the type of logical volume being cache
When a logical volume is cached, it's type become cache and the base
volume is used as device instead of a physical device.
In this case, the scan would fail since the hidden base volume cannot be
accessed directly to read the header.
This redirect the device to the correct physical device underlying the
base volume.